+++ src/frontend/src/modals/IOModal/index.tsx
@@ -1,4 +1,4 @@
-import { useGetMessagesQuery } from "@/controllers/API/queries/messages";
+import { useDeleteMessages, useGetMessagesQuery } from "@/controllers/API/queries/messages";
 import { useQueryClient } from "@tanstack/react-query";
 import { useEffect, useState } from "react";
 import AccordionComponent from "../../components/accordionComponent";
@@ -24,7 +24,6 @@ import { cn } from "../../utils/utils";
 import BaseModal from "../baseModal";
 import IOFieldView from "./components/IOFieldView";
 import SessionView from "./components/SessionView";
-import useRemoveSession from "./components/SessionView/hooks";
 import ChatView from "./components/chatView";
 
 export default function IOModal({
@@ -57,6 +56,32 @@ export default function IOModal({
   );
   const setErrorData = useAlertStore((state) => state.setErrorData);
   const setSuccessData = useAlertStore((state) => state.setSuccessData);
+  const deleteSession = useMessagesStore((state) => state.deleteSession);
+
+
+  const { mutate: deleteSessionFunction } = useDeleteMessages();
+
+
+  function handleDeleteSession(session_id: string) {
+    deleteSessionFunction({
+      ids: messages
+        .filter((msg) => msg.session_id === session_id)
+        .map((msg) => msg.id),
+    },{onSuccess:()=>{
+      setSuccessData({
+        title: "Session deleted successfully.",
+      });
+      deleteSession(session_id);
+    },
+      onError: () => {
+        setErrorData({
+          title: "Error deleting Session.",
+        })
+      }
+    });
+  }
+
+
 
   function startView() {
     if (!chatInput && !chatOutput) {
@@ -125,11 +150,6 @@ export default function IOModal({
       });
     }
   }
-
-  const { handleRemoveSession } = useRemoveSession(
-    setSuccessData,
-    setErrorData,
-  );
 
   useEffect(() => {
     setSelectedTab(inputs.length > 0 ? 1 : outputs.length > 0 ? 2 : 0);
@@ -371,7 +391,7 @@ export default function IOModal({
                               onClick={(e) => {
                                 e.preventDefault();
                                 e.stopPropagation();
-                                handleRemoveSession(session);
+                                handleDeleteSession(session);
                                 if (selectedViewField?.id === session)
                                   setSelectedViewField(undefined);
                               }}
@@ -447,31 +467,31 @@ export default function IOModal({
                     {inputs.some(
                       (input) => input.id === selectedViewField.id,
                     ) && (
-                      <IOFieldView
-                        type={InputOutput.INPUT}
-                        left={false}
-                        fieldType={selectedViewField.type!}
-                        fieldId={selectedViewField.id!}
-                      />
-                    )}
+                        <IOFieldView
+                          type={InputOutput.INPUT}
+                          left={false}
+                          fieldType={selectedViewField.type!}
+                          fieldId={selectedViewField.id!}
+                        />
+                      )}
                     {outputs.some(
                       (output) => output.id === selectedViewField.id,
                     ) && (
-                      <IOFieldView
-                        type={InputOutput.OUTPUT}
-                        left={false}
-                        fieldType={selectedViewField.type!}
-                        fieldId={selectedViewField.id!}
-                      />
-                    )}
+                        <IOFieldView
+                          type={InputOutput.OUTPUT}
+                          left={false}
+                          fieldType={selectedViewField.type!}
+                          fieldId={selectedViewField.id!}
+                        />
+                      )}
                     {sessions.some(
                       (session) => session === selectedViewField.id,
                     ) && (
-                      <SessionView
-                        session={selectedViewField.id}
-                        id={currentFlow!.id}
-                      />
-                    )}
+                        <SessionView
+                          session={selectedViewField.id}
+                          id={currentFlow!.id}
+                        />
+                      )}
                   </div>
                 </div>
               )}
